pyproject.toml
[project]
name = "token-battles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "coincurve>=18.0.0",
    "flask-limiter>=3.9.0",
    "bech32>=1.2.0",
    "boto3>=1.35.0",
    "alembic>=1.13.2",
    "psycopg[binary]>=3.2.1",
    "pillow>=10.0.0",
    "whitenoise>=6.11.0",
    "pymysql>=1.1.2",
    "redis>=5.0.0",
    "rq>=1.15.1",
    "rq-scheduler>=0.13.1",
    "websocket-client>=1.8.0",
    "requests>=2.31.0",
]
